% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_daytime.R
\name{add_daytime}
\alias{add_daytime}
\title{Function to calculate daytimes and add a \code{daytime} variable to a tibble.}
\usage{
add_daytime(df, as_factor = FALSE)
}
\arguments{
\item{df}{Input tibble to be transformed. \code{df} must contain
\code{date}, \code{latitude}, and \code{longitude} variables. The 
\code{"date"} variable must be of \code{POSIXct} class.}

\item{as_factor}{Should the \code{daytime} variable be a factor?}
}
\value{
Tibble, \code{df} with an additional \code{daytime} variable.
}
\description{
The sunrise and sunset calculations are done by \strong{suncalc}'s 
\code{\link[suncalc]{getSunlightTimes}} function.
}
\seealso{
\code{\link[suncalc]{getSunlightTimes}}
}
