% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sunlight_times.R
\name{calculate_sunlight_times}
\alias{calculate_sunlight_times}
\title{Function to calculate sunlight times.}
\usage{
calculate_sunlight_times(df)
}
\arguments{
\item{df}{Input data frame or tibble with \code{date}, \code{latitude}, and 
\code{longitude} variables.}
}
\value{
Tibble.
}
\description{
The calculations are done the \code{\link{getSunlightTimes}} function in the 
\strong{suncalc} package.
}
\examples{

# Create a mapping table with the correctly named inputs for the Vatican City
data_mapping <- data.frame(
  date = Sys.Date(), latitude = 41.90223, longitude = 12.45637
)

# Calculate sunlight time
calculate_sunlight_times(data_mapping)

}
\seealso{
\code{\link{getSunlightTimes}}
}
