% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_file_metadata_coordinates.R
\name{clean_file_metadata_coordinates}
\alias{clean_file_metadata_coordinates}
\title{Function to clean and convert non-decimal degrees, usually from photos' 
metadata strings to to decimal degrees.}
\usage{
clean_file_metadata_coordinates(x, round = 6)
}
\arguments{
\item{x}{Vector containing coordinates in this format: 
\code{53 deg 37\' 56.49\" N}.}

\item{round}{Number of decimal points to round the coordinates to.}
}
\value{
Numeric vector.
}
\description{
Function to clean and convert non-decimal degrees, usually from photos' 
metadata strings to to decimal degrees.
}
\examples{

# An example latitude
clean_file_metadata_coordinates('53 deg 37\' 56.49\" N')

}
\author{
Stuart K. Grange
}
