% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_wkt.R
\name{coords_to_wkt}
\alias{coords_to_wkt}
\title{Function to create WKT geometry strings with string processing functions to 
allow for potentially missing geometries.}
\usage{
coords_to_wkt(x, y, geometry_type = "points", round = NA)
}
\arguments{
\item{x}{\code{x}/longitude vector.}

\item{y}{\code{y}/latitude numeric vector.}

\item{geometry_type}{Type of geometry to create. Only \code{"points"} are 
supported.}

\item{round}{Decimal points to round \code{x} and \code{y} vectors.}
}
\value{
Character vector.
}
\description{
Function to create WKT geometry strings with string processing functions to 
allow for potentially missing geometries.
}
\author{
Stuart K. Grange
}
