% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_constants.R
\name{crs_wgs84}
\alias{crs_wgs84}
\alias{crs_swiss}
\alias{crs_swiss_plus}
\alias{crs_bng}
\alias{crs_nztm}
\alias{crs_gda_94}
\alias{crs_gda_2020}
\alias{crs_with_prefix}
\title{Built-in projection/coordinate reference system constants.}
\usage{
crs_wgs84

crs_swiss

crs_swiss_plus

crs_bng

crs_nztm

crs_gda_94

crs_gda_2020

crs_with_prefix(crs)
}
\arguments{
\item{crs}{Coordinate reference system integer.}
}
\value{
Integer with length of \code{1}.
}
\description{
Built-in projection/coordinate reference system constants.
}
\seealso{
\code{\link[sf]{st_transform}}, \code{\link[terra]{project}}
}
\author{
Stuart K. Grange
}
