% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_constants.R
\docType{data}
\name{crs_wgs_84}
\alias{crs_wgs_84}
\alias{crs_swiss}
\alias{crs_swiss_plus}
\title{Built in projection constants.}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
crs_wgs_84

crs_swiss

crs_swiss_plus
}
\value{
Integer with length of \code{1}.
}
\description{
Built in projection constants.
}
\author{
Stuart K. Grange
}
\keyword{datasets}
