% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_by_haversine.R
\name{distance_by_haversine}
\alias{distance_by_haversine}
\title{Function to calculate distances between two points in metres or kilometres.}
\usage{
distance_by_haversine(
  latitude,
  longitude,
  latitude_lag = NA,
  longitude_lag = NA,
  radius = 6371 * 1000
)
}
\arguments{
\item{latitude}{A vector of latitude values in decimal degrees.}

\item{longitude}{A vector of longitude values in decimal degrees.}

\item{latitude_lag}{A vector of latitude values in decimal degrees.}

\item{longitude_lag}{A vector of longitude values in decimal degrees.}

\item{radius}{The radius of the Earth in kilometres. The default value is 
6 371 000 m (6371 km).}
}
\value{
Numeric vector representing distance in metres
}
\description{
\code{distance_by_haversine} is used when distances between latitude and 
longitude-pairs need to be represented in metre or kilometre units. The 
distances between two points is based on the Haversine Formula which is 
approximate and is most appropriate when used for short-distance calculations.
}
\details{
The Haversine Formula does not compensate for Earth's non-spherical shape, 
will overestimate distances in the polar regions but will underestimate 
distances near the equator. For more information see 
\url{https://en.wikipedia.org/wiki/Haversine_formula}

\code{distance_by_haversine} is not a spatial function and does not have any
geographic library dependencies.
}
\examples{

# What is the distance between Oxford's Radcliffe Camera and Cliveden House 
# in Buckinghamshire? 
distance_by_haversine(51.753447, -1.254024, 51.558170, -0.688250)

}
\author{
Stuart K. Grange
}
