% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_boundary.R
\name{get_osm_boundary}
\alias{get_osm_boundary}
\title{Function to get OpenStreetMap boundaries as \strong{sf} polygons.}
\usage{
get_osm_boundary(id, sleep = 1, verbose = FALSE, progress = FALSE)
}
\arguments{
\item{id}{A vector of OpenStreetMap relations. An integer key.}

\item{sleep}{Number of seconds between server queries. This is useful if 
many large polygons are being requested.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
\strong{sf} polygons.
}
\description{
\code{get_osm_boundary} accesses 
\href{http://polygons.openstreetmap.fr/index.py}{this} polygon creator tool.
}
\examples{

# Get North York Moors National Park boundary
sf_moors <- get_osm_boundary(409150)

}
\author{
Stuart K. Grange
}
