% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_extact.R
\name{ra_extact}
\alias{ra_extact}
\title{Function to extract values from a \strong{terra} \code{SpatRaster} object for
\strong{sf} points.}
\usage{
ra_extact(ra, sf_points, na.rm = FALSE, verbose = FALSE)
}
\arguments{
\item{ra}{A \code{SpatRaster} object.}

\item{sf_points}{A \strong{sf} points object.}

\item{na.rm}{Should missing values in the \code{ra} object be removed?}

\item{verbose}{Should the function give messages?}
}
\description{
Function to extract values from a \strong{terra} \code{SpatRaster} object for
\strong{sf} points.
}
\author{
Stuart K. Grange
}
