% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_interpolate.R
\name{ra_interpolate}
\alias{ra_interpolate}
\title{Function to interpolate a raster object with an inverse distance weighted
method.}
\usage{
ra_interpolate(ra, sf, variable = "value", n_max = Inf, na.rm = FALSE)
}
\arguments{
\item{ra}{A \code{SpatRaster} object.}

\item{sf}{A \strong{sf} object, usually a \strong{sf} points object/}

\item{variable}{Variable name to interpolate in \code{ra}.}

\item{n_max}{The number of nearest observations that should be used for 
interpolation.}

\item{na.rm}{Should \code{NA} be omitted?}
}
\value{
A \code{SpatRaster} object.
}
\description{
Function to interpolate a raster object with an inverse distance weighted
method.
}
\author{
Stuart K. Grange
}
