% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_max.R
\name{ra_max}
\alias{ra_max}
\alias{ra_min}
\title{Functions to find which cells in a \strong{terra} object contains the maximum
and minimum values.}
\usage{
ra_max(ra)

ra_min(ra)
}
\arguments{
\item{ra}{A \strong{terra} \code{SpatRaster} layer.}
}
\value{
\strong{sf} points.
}
\description{
Functions to find which cells in a \strong{terra} object contains the maximum
and minimum values.
}
\author{
Stuart K. Grange
}
