% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_read.R
\name{ra_read}
\alias{ra_read}
\title{Function to read a raster data file as a \strong{terra} \code{SpatRaster} 
object.}
\usage{
ra_read(file, variable = NA, layers = NULL)
}
\arguments{
\item{file}{Raster file name to read.}

\item{variable}{A vector of the variables within \code{file} to be read.}

\item{layers}{An integer or character to select from \code{file}.}
}
\value{
\code{SpatRaster} object.
}
\description{
Function to read a raster data file as a \strong{terra} \code{SpatRaster} 
object.
}
\author{
Stuart K. Grange
}
