% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_read.R
\name{ra_read_nested}
\alias{ra_read_nested}
\title{Function to read a raster file in a nested tibble.}
\usage{
ra_read_nested(
  file,
  separate_layers = FALSE,
  warn = TRUE,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{file}{A vector of raster files name to read.}

\item{separate_layers}{Should all the layers within a variable 
(\strong{terra}'s \code{varname}) be read separately? Not usually desired for
time series, but for some data sources the layers within a variable form 
distinct variables.}

\item{warn}{Should the function raise warnings? Lower-level GDAL warnings can
be raised for myriad reasons, but they are often messages, and are not "true"
warnings and therefore can be suppressed.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Nested tibble grouped by rows.
}
\description{
Function to read a raster file in a nested tibble.
}
\seealso{
\code{\link{ra_read}}
}
\author{
Stuart K. Grange
}
