% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_to_contours.R
\name{ra_to_contours}
\alias{ra_to_contours}
\title{Function to transform a \strong{terra} layer to contour lines.}
\usage{
ra_to_contours(ra, levels = 20)
}
\arguments{
\item{ra}{A \strong{terra} \code{SpatRaster} layer.}

\item{levels}{Number of levels to split the numeric scale of \code{ra} into.}
}
\value{
\strong{sf} lines.
}
\description{
Function to transform a \strong{terra} layer to contour lines.
}
\author{
Stuart K. Grange
}
