% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_to_df.R
\name{ra_to_df}
\alias{ra_to_df}
\title{Function to transform a raster object to a tibble.}
\usage{
ra_to_df(ra)
}
\arguments{
\item{ra}{A \strong{terra} \code{SpatRaster} object.}
}
\value{
Tibble.
}
\description{
Function to transform a raster object to a tibble.
}
\author{
Stuart K. Grange
}
