% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_transform.R
\name{ra_transform}
\alias{ra_transform}
\title{Function to project a \strong{terra} raster object's CRS projection system.}
\usage{
ra_transform(ra, crs)
}
\arguments{
\item{ra}{A \strong{terra} raster object.}

\item{crs}{A projection/coordinate reference system (CRS) integer.}
}
\value{
A \strong{terra} raster object.
}
\description{
Function to project a \strong{terra} raster object's CRS projection system.
}
\seealso{
\code{\link{project}}, \code{\link{st_transform}}, 
\code{\link{crs_wgs84}}
}
