% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_gpx}
\alias{read_gpx}
\title{Function to read \code{GPX} files as tabular data.}
\usage{
read_gpx(
  file,
  transform = TRUE,
  creator = FALSE,
  names_to = rlang::zap(),
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{file}{A vector of GPS files.}

\item{transform}{Should latitude and longitude be used to calculate speed and
distance if the file contains the appropriate variables?}

\item{creator}{Should the \code{creator} variable be extracted from the file?}

\item{names_to}{An optional variable name for file name in the returned 
tibble.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Tibble.
}
\description{
Function to read \code{GPX} files as tabular data.
}
\author{
Stuart K. Grange
}
