% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_add_id.R
\name{sf_add_id}
\alias{sf_add_id}
\title{Function to add an integer ID to a \code{sf} object.}
\usage{
sf_add_id(sf, name = "id_sf")
}
\arguments{
\item{sf}{A \code{sf} object.}

\item{name}{The name of the ID variable.}
}
\value{
\code{sf} with an additional ID variable.
}
\description{
Function to add an integer ID to a \code{sf} object.
}
\author{
Stuart K. Grange
}
