% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_buffer_and_smooth.R
\name{sf_buffer_and_smooth}
\alias{sf_buffer_and_smooth}
\title{Function to buffer a \strong{sf} object and subsequently smooth the buffered
output.}
\usage{
sf_buffer_and_smooth(sf, buffer = 1000, smoothness = 15, tolerance = 0.1)
}
\arguments{
\item{sf}{A \strong{sf} object.}

\item{buffer}{Buffer distance. See \code{\link[sf:geos_unary]{st_buffer}} for 
details.}

\item{smoothness}{The smoothness and level of generalisation parameter.}

\item{tolerance}{The tolerance parameter.}
}
\value{
\strong{sf} buffered and smoothed.
}
\description{
Function to buffer a \strong{sf} object and subsequently smooth the buffered
output.
}
\author{
Stuart K. Grange
}
