% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_class.R
\name{sf_class}
\alias{sf_class}
\title{Function to get the class of a \strong{sf} spatial object.}
\usage{
sf_class(sf, as_lower = TRUE)
}
\arguments{
\item{sf}{A \strong{sf} spatial object.}

\item{as_lower}{Should the return be lower-case?}
}
\value{
Character vector with length of \code{1}.
}
\description{
Function to get the class of a \strong{sf} spatial object.
}
\author{
Stuart K. Grange
}
