% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_create_bounding_box.R
\name{sf_create_bounding_box}
\alias{sf_create_bounding_box}
\title{Function to create a \code{sf} bounding box for a \code{sf} object.}
\usage{
sf_create_bounding_box(sf, buffer = NA)
}
\arguments{
\item{sf}{A \code{sf} object.}

\item{buffer}{An optional buffer to apply before calculating the bounding
box.}
}
\value{
A \code{sf} polygon.
}
\description{
Function to create a \code{sf} bounding box for a \code{sf} object.
}
\author{
Stuart K. Grange
}
