% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_create_mesh.R
\name{sf_create_mesh}
\alias{sf_create_mesh}
\title{Function to create a mesh or grid of polygons to cover an sf object, 
typically a larger polygon.}
\usage{
sf_create_mesh(
  sf,
  cellsize = NA,
  n = 10,
  hexagonal = TRUE,
  filter = FALSE,
  intersect = FALSE
)
}
\arguments{
\item{sf}{Input \code{sf} object, typically a polygon.}

\item{cellsize}{Numeric vector of length \code{1} or \code{2} with the target 
cellsize.}

\item{n}{Integer vector of length \code{1} or \code{2}, number of grid cells
in the \code{x} or \code{y} directions.}

\item{hexagonal}{Should a hexagonal (not square) mesh be made?}

\item{filter}{Should the mesh be filtered to \code{sf}?}

\item{intersect}{Should the cropped/clipped to \code{sf}?}
}
\value{
A \code{sf} object of polygon geometry type.
}
\description{
Function to create a mesh or grid of polygons to cover an sf object, 
typically a larger polygon.
}
\seealso{
\code{\link[sf]{st_make_grid}}
}
