% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_extreme_points.R
\name{sf_extreme_points}
\alias{sf_extreme_points}
\title{Function to get extreme points for a \code{sf} object.}
\usage{
sf_extreme_points(sf)
}
\arguments{
\item{sf}{A \code{sf} object.}
}
\value{
\code{sf} points.
}
\description{
Function to get extreme points for a \code{sf} object.
}
\author{
Stuart K. Grange
}
