% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_from_df.R
\name{sf_from_df}
\alias{sf_from_df}
\title{Function to promote a data frame or tibble to a \strong{sf} spatial object.}
\usage{
sf_from_df(
  df,
  geometry_type = "points",
  coords = c("longitude", "latitude"),
  by = NULL,
  crs = sf::st_crs("wgs84"),
  geometry_name = "geometry",
  verbose = TRUE
)
}
\arguments{
\item{df}{Data frame or tibble containing spatial data to be promoted to a 
\strong{sf} spatial object.}

\item{geometry_type}{Type of geometry to promote data to. One of \code{"points"},
\code{"lines"}, or \code{"polygons"}.}

\item{coords}{What variables in \code{df} contain the \code{x} and \code{y}
coordinates (in that order).}

\item{by}{What variable(s) should be used as a grouping variable?}

\item{crs}{What coordinate reference system should are \code{coords} stored
in?}

\item{geometry_name}{What variable name should the geometry variable take?}

\item{verbose}{Should the function give messages?}
}
\value{
A \strong{sf} spatial object.
}
\description{
Function to promote a data frame or tibble to a \strong{sf} spatial object.
}
\author{
Stuart K. Grange
}
