% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_start_point.R
\name{sf_start_point}
\alias{sf_start_point}
\alias{sf_end_point}
\title{Function to return the start and end points of a \strong{sf} line object.}
\usage{
sf_start_point(sf)

sf_end_point(sf)
}
\arguments{
\item{sf}{A \strong{sf} line object.}
}
\value{
\strong{sf} point object.
}
\description{
Function to return the start and end points of a \strong{sf} line object.
}
\author{
Stuart K. Grange
}
