% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_df.R
\name{sf_to_df}
\alias{sf_to_df}
\title{Function to transform a \strong{sf} spatial object to a "long" data 
frame/tibble.}
\usage{
sf_to_df(sf, rename = FALSE)
}
\arguments{
\item{sf}{A \strong{sf} spatial object.}

\item{rename}{Should the generic \code{"x"} and \code{"y"} names be renamed
to \code{"longitude"} and \code{"latitude"}?}
}
\value{
Tibble.
}
\description{
Function to transform a \strong{sf} spatial object to a "long" data 
frame/tibble.
}
\author{
Stuart K. Grange
}
