% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_sf_functions.R
\name{st_distance_single}
\alias{st_distance_single}
\title{Function to calculate spatial distance between an \code{sf} object with a 
single geometry and a second \code{sf} object with many geometries.}
\usage{
st_distance_single(sf, sf_single, as_integer = FALSE)
}
\arguments{
\item{sf}{A \code{sf} object with many geometries.}

\item{sf_single}{A \code{sf} object with a single geometry.}

\item{as_integer}{Should the return be an integer vector?}
}
\value{
Numeric or integer vector with the length of \code{sf}.
}
\description{
Function to calculate spatial distance between an \code{sf} object with a 
single geometry and a second \code{sf} object with many geometries.
}
\author{
Stuart K. Grange
}
