% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_erase.R
\name{st_erase}
\alias{st_erase}
\title{Function to use a \strong{sf} object to erase pieces of another \strong{sf}
object.}
\usage{
st_erase(sf_x, sf_y)
}
\arguments{
\item{sf_x}{First \strong{sf} object.}

\item{sf_y}{Second \strong{sf} object.}
}
\value{
An \strong{sf}.
}
\description{
Function to use a \strong{sf} object to erase pieces of another \strong{sf}
object.
}
\author{
Stuart K. Grange
}
