% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_intersection_test.R
\name{st_intersection_test}
\alias{st_intersection_test}
\title{Function to test if \code{sf} objects intersect with one another.}
\usage{
st_intersection_test(x, y)
}
\arguments{
\item{x, y}{\code{sf} objects to test. \code{st_intersection_test} will test 
if the features in \code{x} intersect with any feature in \code{y}.}
}
\value{
Logical vector the length of \code{x}.
}
\description{
Function to test if \code{sf} objects intersect with one another.
}
\author{
Stuart K. Grange
}
