% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_sf_functions.R
\name{st_length_simple}
\alias{st_length_simple}
\title{Function to calculate lengths of a \code{sf} object and return a vector.}
\usage{
st_length_simple(sf, as_integer = FALSE)
}
\arguments{
\item{sf}{\code{sf} object.}

\item{as_integer}{Should the return be an integer vector?}
}
\value{
Numeric or integer vector.
}
\description{
Function to calculate lengths of a \code{sf} object and return a vector.
}
\author{
Stuart K. Grange
}
