% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_overlaps_test.R
\name{st_overlaps_test}
\alias{st_overlaps_test}
\title{Function to test if \code{sf} objects overlap with one another.}
\usage{
st_overlaps_test(sf, any = FALSE)
}
\arguments{
\item{sf}{An \code{sf} object to test for overlaps.}

\item{any}{Should \code{\link{any}} be use to test if some values are 
\code{TRUE}?}
}
\value{
Logical vector with the length of \code{sf} or \code{1}.
}
\description{
Function to test if \code{sf} objects overlap with one another.
}
\author{
Stuart K. Grange
}
