% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geo_tiff.R
\name{write_geo_tiff}
\alias{write_geo_tiff}
\alias{write_esri_grid_ascii}
\title{Functions to export raster objects.}
\usage{
write_geo_tiff(ra, file)

write_esri_grid_ascii(ra, file)
}
\arguments{
\item{ra}{Raster object to export.}

\item{file}{File to export raster object too.}
}
\value{
Invisible, \code{file}.
}
\description{
Functions to export raster objects.
}
\examples{
\dontrun{

# Export raster object as geo tiff 
write_geo_tiff(ra, "raster_object.tif")

# Or text file
write_esri_grid_ascii(ra, "raster_object.grd")

}

}
\author{
Stuart K. Grange
}
