% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_shapefile.R
\name{write_shapefile}
\alias{write_shapefile}
\title{Function to write a \strong{sf} spatial object to shapefiles.}
\usage{
write_shapefile(sf, file)
}
\arguments{
\item{sf}{A \strong{sf} spatial object.}

\item{file}{File name to export.}
}
\value{
Invisible \code{file}.
}
\description{
Function to write a \strong{sf} spatial object to shapefiles.
}
\author{
Stuart K. Grange
}
