% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_constants.R
\docType{data}
\name{crs_wgs84}
\alias{crs_wgs84}
\alias{crs_swiss}
\alias{crs_swiss_plus}
\alias{crs_bng}
\alias{crs_nztm}
\alias{crs_with_prefix}
\title{Built-in projection/coordinate reference system constants.}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
crs_wgs84

crs_swiss

crs_swiss_plus

crs_bng

crs_nztm

crs_with_prefix(crs)
}
\arguments{
\item{crs}{Coordinate reference system integer.}
}
\value{
Integer with length of \code{1}.
}
\description{
Built-in projection/coordinate reference system constants.
}
\seealso{
[func(st_crs)]
}
\author{
Stuart K. Grange
}
\keyword{datasets}
