% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_extract.R
\name{ra_extract}
\alias{ra_extract}
\title{Function to extract values from a \strong{terra} \code{SpatRaster} object for
\strong{sf} points.}
\usage{
ra_extract(
  ra,
  sf,
  drop_ids = FALSE,
  warn = TRUE,
  na.rm = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ra}{A \code{SpatRaster} object.}

\item{sf}{A \strong{sf} object. \code{sf} can be points, lines, or polygons.}

\item{drop_ids}{Should identifiers from the raster object (\code{cell_number} 
and \code{layer}) be dropped from the return?}

\item{warn}{Should the function raise warnings? Lower-level GDAL warnings can
be raised for myriad reasons, but they are often messages, and are not "true"
warnings and therefore can be suppressed.}

\item{na.rm}{Should missing values in the \code{ra} object be removed?}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to extract values from a \strong{terra} \code{SpatRaster} object for
\strong{sf} points.
}
\seealso{
\code{\link{ra_read}}, \code{\link{ra_extract_nested}}
}
\author{
Stuart K. Grange
}
