% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_extract_nested.R
\name{ra_extract_nested}
\alias{ra_extract_nested}
\title{Function to extract values from raster (\code{SpatRaster}) objects when 
contained in a nested tibble.}
\usage{
ra_extract_nested(
  df_nest,
  sf_points,
  drop_ids = TRUE,
  na.rm = FALSE,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{df_nest}{A nested tibble from \code{\link{ra_read_nested}}.}

\item{sf_points}{A \strong{sf} points object.}

\item{drop_ids}{Should identifiers from the raster object (\code{cell_number} 
and \code{layer}) be dropped from the return?}

\item{na.rm}{Should missing values in the \code{ra} object be removed?}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
A nested tibble, \code{df_nest} with an additional \code{extract} 
variable.
}
\description{
Function to extract values from raster (\code{SpatRaster}) objects when 
contained in a nested tibble.
}
\seealso{
\code{\link{ra_read_nested}}, \code{\link{ra_extract}}
}
\author{
Stuart K. Grange
}
