% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_read.R
\name{ra_read_nested}
\alias{ra_read_nested}
\title{Function to read a raster file in a nested tibble.}
\usage{
ra_read_nested(file)
}
\arguments{
\item{file}{Raster file name to read.}
}
\value{
Nested tibble grouped by rows.
}
\description{
Function to read a raster file in a nested tibble.
}
\seealso{
[ra_read()]
}
\author{
Stuart K. Grange
}
