% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_read.R
\name{ra_read_nested}
\alias{ra_read_nested}
\title{Function to read a raster file in a nested tibble.}
\usage{
ra_read_nested(file, verbose = FALSE, progress = FALSE)
}
\arguments{
\item{file}{A vector of raster files name to read.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Nested tibble grouped by rows.
}
\description{
Function to read a raster file in a nested tibble.
}
\seealso{
\code{\link{ra_read}}
}
\author{
Stuart K. Grange
}
