% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gpx.R
\name{write_gpx}
\alias{write_gpx}
\title{Function to plot spatial objects with an interactive \strong{leaflet} plot.}
\usage{
write_gpx(sf, file, transform = FALSE)
}
\arguments{
\item{sf}{A \strong{sf} spatial object.}

\item{file}{File name to export.}

\item{transform}{Should the projection system be transformed to WGS84 before
plotting?}
}
\value{
Invisible \code{file}.
}
\description{
Function to plot spatial objects with an interactive \strong{leaflet} plot.
}
\author{
Stuart K. Grange
}
